[REFACTOR] exception 패키지 위치를 변경하였습니다.
