+++ src/main/java/techcourse/myblog/exception/InvalidDataFormException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.exception;
+package techcourse.myblog.exception;
 
 public class InvalidDataFormException extends IllegalArgumentException {
     public InvalidDataFormException(String message) {
