+++ src/main/java/techcourse/myblog/exception/NotFindUserEmailException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.exception;
+package techcourse.myblog.exception;
 
 public class NotFindUserEmailException extends IllegalArgumentException {
     public NotFindUserEmailException() {
