+++ src/main/java/techcourse/myblog/exception/UnequalPasswordException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.exception;
+package techcourse.myblog.exception;
 
 public class UnequalPasswordException extends IllegalArgumentException {
     public UnequalPasswordException() {
